+++ tests/consensus_tests/test_strict_mode.py
@@ -1,7 +1,7 @@
 import logging
 import pathlib
 
-from .fixtures import create_collection, upsert_random_points, upsert_points, random_dense_vector, drop_collection
+from .fixtures import create_collection, upsert_random_points, upsert_points, random_dense_vector, set_strict_mode
 from .utils import *
 
 logging.basicConfig(level=logging.DEBUG)
@@ -11,6 +11,8 @@ N_PEERS = 2
 N_SHARDS = 1
 N_REPLICAS = 1
 COLLECTION_NAME = "test_collection_strict_mode"
+
+UPSERT_REQUEST_DELAY = 0.3
 
 
 def test_strict_mode_upsert(tmp_path: pathlib.Path):
@@ -43,7 +45,7 @@ def test_strict_mode_upsert(tmp_path: pathlib.Path):
 def test_strict_mode_upsert_no_local_shard(tmp_path: pathlib.Path):
     peer_urls, peer_dirs, bootstrap_url = start_cluster(tmp_path, N_PEERS)
 
-    create_collection(peer_urls[0], collection=COLLECTION_NAME, shard_number=N_SHARDS, replication_factor=N_REPLICAS, sharding_method="custom")
+    create_collection(peer_urls[0], collection=COLLECTION_NAME, shard_number=1, replication_factor=N_REPLICAS, sharding_method="custom")
 
     wait_collection_exists_and_active_on_all_peers(collection_name=COLLECTION_NAME, peer_api_uris=peer_urls)
 
@@ -60,23 +62,26 @@ def test_strict_mode_upsert_no_local_shard(tmp_path: pathlib.Path):
     for _ in range(32):
         point = {"id": 1, "payload": {}, "vector": random_dense_vector()}
         upsert_points(peer_urls[0], [point], collection_name=COLLECTION_NAME, shard_key="non_leader").raise_for_status()
+        time.sleep(UPSERT_REQUEST_DELAY)  # Give qdrant some time to update shard statistics
 
     set_strict_mode(peer_urls[0], COLLECTION_NAME, {
         "enabled": True,
-        "max_collection_vector_size_bytes": 30,
+        "max_collection_vector_size_bytes": 33,
     })
 
     for _ in range(32):
-        point = {"id": 1, "payload": {}, "vector": random_dense_vector()}
+        point = {"id": 2, "payload": {}, "vector": random_dense_vector()}
         upsert_points(peer_urls[0], [point], collection_name=COLLECTION_NAME, shard_key="non_leader").raise_for_status()
+        time.sleep(0.3)
 
     for _ in range(32):
-        point = {"id": 2, "payload": {}, "vector": random_dense_vector()}
+        point = {"id": 3, "payload": {}, "vector": random_dense_vector()}
         res = upsert_points(peer_urls[0], [point], collection_name=COLLECTION_NAME, shard_key="non_leader")
         if not res.ok:
             assert "Max vector storage size" in res.json()['status']['error']
             assert not res.ok
             return
+        time.sleep(UPSERT_REQUEST_DELAY)
 
     assert False, "Should have blocked upsert but didn't"
 
@@ -91,31 +96,26 @@ def test_strict_mode_upsert_local_shard(tmp_path: pathlib.Path):
     for _ in range(32):
         point = {"id": 1, "payload": {}, "vector": random_dense_vector()}
         upsert_points(peer_urls[0], [point], collection_name=COLLECTION_NAME).raise_for_status()
+        time.sleep(UPSERT_REQUEST_DELAY)  # Give qdrant some time to update shard statistics
 
     set_strict_mode(peer_urls[0], COLLECTION_NAME, {
         "enabled": True,
-        "max_collection_vector_size_bytes": 30,
+        "max_collection_vector_size_bytes": 33,
     })
 
     for _ in range(32):
-        point = {"id": 1, "payload": {}, "vector": random_dense_vector()}
+        point = {"id": 2, "payload": {}, "vector": random_dense_vector()}
         upsert_points(peer_urls[0], [point], collection_name=COLLECTION_NAME).raise_for_status()
+        time.sleep(UPSERT_REQUEST_DELAY)
 
     for _ in range(32):
-        point = {"id": 2, "payload": {}, "vector": random_dense_vector()}
+        point = {"id": 3, "payload": {}, "vector": random_dense_vector()}
         res = upsert_points(peer_urls[0], [point], collection_name=COLLECTION_NAME)
         if not res.ok:
             assert "Max vector storage size" in res.json()['status']['error']
             assert not res.ok
             return
+        time.sleep(UPSERT_REQUEST_DELAY)
 
     assert False, "Should have blocked upsert but didn't"
 
-
-def set_strict_mode(peer_id, collection_name, strict_mode_config):
-    requests.patch(
-        f"{peer_id}/collections/{collection_name}",
-        json={
-            "strict_mode_config": strict_mode_config,
-        },
-    ).raise_for_status()
